mv launch_server
